refactor(network): move code
